KDW003: Fix bug #119386

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/find/application/timeleaveapplication/dto/TimeLeaveRemaining.java
@@ -56,7 +56,7 @@ public class TimeLeaveRemaining {
     private String grantDate;
 
     // 付与日数
-    private Integer grantedDays;
+    private Double grantedDays;
 
     public static TimeVacationRemainingOutput setDataOutput(TimeLeaveRemaining dto) {
         return new TimeVacationRemainingOutput(
